altos: Switch LED pin number to 8 bits
This assumes each port has fewer than 256 pins and avoids some
compiler warnings

+++ src/drivers/ao_led.c
@@ -20,7 +20,7 @@
 
 static const struct {
 	void		*port;
-	uint16_t	pin;
+	uint8_t		pin;
 } ao_leds[] = {
 #ifdef LED_0_PORT
     [0] { LED_0_PORT, LED_0_PIN },
